moved mutation tracker into its own location
